requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

httpx==0.27.2
PyNaCl==1.5.0
python-socketio==5.11.4
rich==13.9.4
